PCTO_NI/pipeman-master/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.1
decorator==4.4.2
Deprecated==1.2.14
Django==4.2.1
django-betterforms==2.0.0
django-crispy-forms==2.0
django-filter==23.2
django-impersonate==1.9.1
django-oauth-toolkit==2.3.0
django-unicorn==0.51.1
django-vuexy==1.0.7
djangorestframework==3.14.0
fontawesomefree==6.4.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
jwcrypto==1.5.0
Markdown==3.4.3
mysqlclient==2.1.1
oauthlib==3.2.2
orjson==3.9.0
Pillow==9.5.0
pycparser==2.21
python-gitlab==3.14.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
shortuuid==1.0.11
smmap==5.0.0
sorl-thumbnail==2023.3.10
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
wrapt==1.15.0
